pyproject.toml
[tool.poetry]
name = "synthea-api"
version = "0.1.0"
description = ""
authors = ["Javier Salvador <javier@jsrf.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
requests = "^2.31.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
sqlalchemy = "^2.0.23"
pydantic = "^2.5.2"
isort = "^5.12.0"
httpx = "^0.25.2"
psycopg2-binary = "^2.9.9"
bs4 = "^0.0.1"
python-multipart = "^0.0.6"
pypdf2 = "^3.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-env = "^1.1.3"
factory-boy = "^3.3.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
env = [
    "HOME=~/tmp",
    "RUN_ENV=test",
    "SQLALCHEMY_DATABASE_URL=sqlite:///test.db"
]